Initial commit for motivational quotes email send
This project gives an overview of how we can send motivational quotes via email to the recipients. It uses smtplib to send email, pandas to get the email information from the csv files. This code can be run on cloud say onto pythonanywhere.com and can be scheduled to run everyday